# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/atleta": {
            "post": {
                "description": "Cadastra um novo atleta no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Atletas"
                ],
                "summary": "Cria um novo atleta",
                "parameters": [
                    {
                        "description": "Dados do atleta",
                        "name": "atleta",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Atleta"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Atleta"
                        }
                    },
                    "400": {
                        "description": "Falha ao cadastrar atleta",
                        "schema": {
                            "$ref": "#/definitions/models.Atleta"
                        }
                    }
                }
            }
        },
        "/atleta/{cpf}": {
            "get": {
                "description": "Busca um atleta pelo CPF informado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Atletas"
                ],
                "summary": "Exibe um atleta por CPF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CPF do atleta",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Atleta"
                        }
                    },
                    "404": {
                        "description": "Atleta não encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.Atleta"
                        }
                    }
                }
            },
            "patch": {
                "description": "Atualiza os dados de um atleta existente pelo CPF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Atletas"
                ],
                "summary": "Atualiza um atleta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CPF do atleta",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados do atleta",
                        "name": "atleta",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Atleta"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Atleta"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos para atualização",
                        "schema": {
                            "$ref": "#/definitions/models.Atleta"
                        }
                    },
                    "404": {
                        "description": "Atleta não encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.Atleta"
                        }
                    }
                }
            }
        },
...
        "/equipe": {
            "post": {
                "description": "Cadastra uma nova equipe no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipes"
                ],
                "summary": "Cria uma nova equipe",
                "parameters": [
                    {
                        "description": "Dados da equipe",
                        "name": "equipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Equipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Equipe"
                        }
                    },
                    "400": {
                        "description": "Falha ao cadastrar equipe",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Falha ao salvar equipe no banco de dados",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/equipe/{id}": {
            "get": {
                "description": "Busca uma equipe pelo ID informado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipes"
                ],
                "summary": "Exibe uma equipe por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da equipe",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Equipe"
                        }
                    },
                    "404": {
                        "description": "Equipe não encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove uma equipe do sistema pelo ID informado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipes"
                ],
                "summary": "Deleta uma equipe por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da equipe",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipe deletada com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Equipe não encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Falha ao deletar a equipe",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "description": "Atualiza os dados de uma equipe existente pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipes"
                ],
                "summary": "Atualiza uma equipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da equipe",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados da equipe",
                        "name": "equipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Equipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Equipe"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos para atualização",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Equipe não encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Falha ao atualizar a equipe",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Atleta": {
            "description": "Representa um atleta com informações básicas.",
            "type": "object",
            "properties": {
                "ativo": {
                    "description": "Status ativo do atleta",
                    "type": "integer"
                },
                "cpf": {
                    "description": "CPF do atleta",
                    "type": "string"
                },
                "cpf_lider": {
                    "description": "CPF do líder associado",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Data de criação",
                    "type": "string"
                },
                "deletedAt": {
                    "description": "Data de exclusão (opcional)",
                    "type": "string"
                },
                "email": {
                    "description": "Email do atleta",
                    "type": "string"
                },
                "equipe": {
                    "description": "Relacionamento com o modelo Equipe",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Equipe"
                        }
                    ]
                },
                "funcao": {
                    "description": "Função do atleta",
                    "type": "string"
                },
                "id": {
                    "description": "ID do atleta",
                    "type": "integer"
                },
                "id_equipe": {
                    "description": "ID da equipe associada",
                    "type": "integer"
                },
                "idade": {
                    "description": "Idade do atleta",
                    "type": "integer"
                },
                "lider": {
                    "description": "Relacionamento com o modelo Lider",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Lider"
                        }
                    ]
                },
                "nome": {
                    "description": "Nome do atleta",
                    "type": "string"
                },
                "senha": {
                    "description": "Senha do atleta",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Data de atualização",
                    "type": "string"
                }
            }
        },
        "models.Equipe": {
            "description": "Representa uma equipe com informações básicas.",
            "type": "object",
            "properties": {
                "ativo": {
                    "description": "Status ativo da equipe",
                    "type": "integer"
                },
                "createdAt": {
                    "description": "Data de criação",
                    "type": "string"
                },
                "deletedAt": {
                    "description": "Data de exclusão (opcional)",
                    "type": "string"
                },
                "id": {
                    "description": "ID da equipe",
                    "type": "integer"
                },
                "nome": {
                    "description": "Nome da equipe",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Data de atualização",
                    "type": "string"
                }
            }
        },
        "models.Lider": {
            "description": "Representa um líder com informações básicas.",
            "type": "object",
            "properties": {
                "cpf": {
                    "description": "CPF do líder",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Data de criação",
                    "type": "string"
                },
                "deletedAt": {
                    "description": "Data de exclusão (opcional)",
                    "type": "string"
                },
                "email": {
                    "description": "Email do líder",
                    "type": "string"
                },
                "nome": {
                    "description": "Nome do líder",
                    "type": "string"
                },
                "senha": {
                    "description": "Senha do líder",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Data de atualização",
                    "type": "string"
                }
            }
        }
    }
}